API

MacroUtilities.@parse_kwargs — Macro

@parse_kwargs [args] kwarg_spec

Parses the set of keyword expressions given in args according to a series of kwarg_spec specification and sets the corresponding keys in the calling scope.

kwarg_spec must be a block Expr, with each line consisting of an expression of the form

    key = (expected_type = T)

which specifies that for a key-value expression in args of the form key = value, value must have type T.

An alternative form is

    key = (expected_types = (T1, T2, ..., Tn))

which specifies that a key-value expression in args with key = value must have typeof(value) in (T1, T2, ..., Tn)

If the default key is provided, e.g.,

    key = (expected_types = (T1, T2, ..., Tn), default=default_value)

then key is set to default_value if args do not contain a key = value expression.

If the default key is not provided, then args must contain a key = value expression or an ArgumentError will be thrown.

MacroUtilities.from_expr — Method

from_expr(::Type{T}, expr; throw_error::Bool=false, kwargs...)

Parses an expr into an object of type T

The provided kwargs are passed to the underlying parsing function.

If expr cannot be parsed into an object of type T - if throw_error == true, throws an ArgumentError - otherwise, returns nothing

====================

from_expr(::Type{Vector{T}}, expr::Expr; throw_error::Bool=false)

Parses a tuple or vect expr as a Vector{T}. Each argument of expr must be of type T.

MacroUtilities.to_expr — Function

to_expr(x) -> Expr

Converts x to an Expr representation

MacroUtilities.FuncArg — Type

FuncArg(; name, type, value::Any, is_splat::Bool)

Matches a function argument expression

Arguments

name::Union{NotProvided, Symbol} = not_provided: Name of the argument

type::Union{NotProvided, Symbol, Expr} = not_provided: Annotated type of the argument. Only valid when value is not provided.

value::Any = not_provided: Value of the argument. If name is provided, this corresponds to the default value of name in a method definition. Otherwise, this is the value passed as a function argument

is_splat::Bool = false: true if this argument is splatted, false otherwise

MacroUtilities.FuncCall — Type

FuncCall(; funcname::Union{NotProvided, Symbol, Expr}, args::Vector{FuncArg}, kwargs::OrderedDict{Symbol, FuncArg})

Matches a function call expression

MacroUtilities.FuncDef — Type

FuncDef(; header::FuncCall, head::Symbol, whereparams::Any, return_type::Any, body::Any, line::Union{LineNumberNode, NotProvided}, doc::Union{LineNumberNode, NotProvided})

Matches a function definition expression

MacroUtilities.KVExpr — Type

KVExpr(; key, value)

Matches expressions of the form key::Symbol = value

MacroUtilities.is_not_provided — Function

is_not_provided(x) -> Bool

Returns true if the field corresponding to x is not provided, false otherwise.

MacroUtilities.NotProvided — Type

NotProvided

Placeholder type to represent the absence of a field.