API

General Utilities

MacroUtilities.@assert_type — Macro

@assert_type var types

Ensures that typeof(var) is in the list of provided types.

types can be either a single type or a tuple or vect expression of types

Keyword Arg Parsing

MacroUtilities.@parse_kwargs — Macro

@parse_kwargs [args] kwarg_spec

Parses the set of keyword expressions given in args according to a series of kwarg_spec specification and sets the corresponding keys in the calling scope.

kwarg_spec must be a block Expr, with each line consisting of an expression of the form

    key = (expected_type = T)

which specifies that for a key-value expression in args of the form key = value, value must have type T.

An alternative form is

    key = (expected_types = (T1, T2, ..., Tn))

which specifies that a key-value expression in args with key = value must have typeof(value) in (T1, T2, ..., Tn)

If the default key is provided, e.g.,

    key = (expected_types = (T1, T2, ..., Tn), default=default_value)

then key is set to default_value if args do not contain a key = value expression.

If the default key is not provided, then args must contain a key = value expression or an ArgumentError will be thrown.

Syntax Parsing

General

MacroUtilities.from_expr — Function

from_expr(::Type{T}, expr; throw_error::Bool=false, kwargs...)

Parses an expr into an object of type T

The provided kwargs are passed to the underlying parsing function.

If expr cannot be parsed into an object of type T - if throw_error == true, throws an ArgumentError - otherwise, returns nothing

====================

from_expr(::Type{Vector{T}}, expr::Expr; throw_error::Bool=false)

Parses a tuple or vect expr as a Vector{T}. Each argument of expr must be of type T

====================

from_expr(::Type{Vector{T}}, input::T; throw_error::Bool=false)

Returns a singleton Vector{T} containing input

MacroUtilities.to_expr — Function

to_expr(x) -> Expr

Converts x to an Expr representation

MacroUtilities.NotProvided — Type

NotProvided

Placeholder type to represent the absence of a field.

MacroUtilities.is_not_provided — Function

is_not_provided(x) -> Bool

Returns true if the field corresponding to x is not provided, false otherwise.

Expression Types

MacroUtilities.AssignExpr — Type

AssignExpr(; lhs, rhs, allow_kw)

Matches a lhs = rhs assignment expression.

If allow_key == true, also matches an assignment expression with head :kw

MacroUtilities.BlockExpr — Type

BlockExpr(; args)

Matches a :block expression

MacroUtilities.KVExpr — Type

KVExpr(; key, value)

Matches expressions of the form key::Symbol = value

Macro Expressions

MacroUtilities.MacroCall — Type

MacroCall(; name::Union{Symbol, Expr, GlobalRef}, line::Union{LineNumberNode, NotProvided} = not_provided, args::Vector{Any} = [])

Matches a macro call expression. A MacroCall object can be applied to one or more expressions, yielding another MacroCall.

MacroUtilities.doc_macro — Constant

doc_macro

A MacroCall corresponding to Core.@doc

MacroUtilities.assume_effects — Constant

assume_effects

A MacroCall corresponding to Base.@assume_effects

MacroUtilities.assume_foldable — Constant

assume_foldable

A MacroCall corresponding to Base.@assume_effects :foldable

Function Expressions

MacroUtilities.FuncDef — Type

FuncDef(; header::FuncCall, head::Symbol, whereparams::Any, return_type::Any, body::Any, line::Union{LineNumberNode, NotProvided}, doc::Union{LineNumberNode, NotProvided})

Matches a function definition expression

MacroUtilities.FuncDef — Method

FuncDef(f::FuncDef; [header, head, whereparams, return_type, body, line, doc)

Returns a new copy of f, with optional header, head, whereparams, return_type, body, line, or doc values overridden by the keyword arguments.

MacroUtilities.FuncCall — Type

FuncCall(; funcname::Union{NotProvided, Symbol, Expr}, args::Vector{FuncArg}, kwargs::OrderedDict{Symbol, FuncArg})

Matches a function call expression

MacroUtilities.FuncCall — Method

FuncCall(f::FuncCall; [funcname, args, kwargs])

Returns a new copy of f, with optional funcname, args, or kwargs overridden by the keyword argumnets.

MacroUtilities.FuncArg — Type

FuncArg(; name, type, value::Any, is_splat::Bool)

Matches a function argument expression

Arguments

name::Union{NotProvided, Symbol} = not_provided: Name of the argument

type::Union{NotProvided, Symbol, Expr} = not_provided: Annotated type of the argument. Only valid when value is not provided.

value::Any = not_provided: Value of the argument. If name is provided, this corresponds to the default value of name in a method definition. Otherwise, this is the value passed as a function argument

is_splat::Bool = false: true if this argument is splatted, false otherwise

MacroUtilities.FuncArg — Method

FuncArg(f::FuncArg; [name, type, value, is_splat])

Returns a new copy of f, with optional name, type, value, or is_splat overridden by the keyword argumnets.

MacroUtilities.map_args — Function

map_args(f, expr::FuncCall) -> FuncCall
map_args(f, expr::FuncDef) -> FuncCall

Transform the expr by applying f(FuncArg) -> FuncArg to each of its arguments

MacroUtilities.map_kwargs — Function

map_kwargs(f, expr::FuncCall) -> FuncCall
map_kwargs(f, expr::FuncDef) -> FuncCall

Transform the expr by applying f(FuncArg) -> FuncArg to each of its keyword arguments